Add RecordSelectionFormula property and update ProcessReport
Added a new property `RecordSelectionFormula` to the `Data` class in
the `Majorsilence.CrystalCmd.Common` namespace. Updated the
`ProcessReport` method in the `CrystalDocumentWrapper` class within
the `Majorsilence.CrystalCmd.Server.Common` namespace to utilize this
new property. The method now checks if `RecordSelectionFormula` is
not null or whitespace and sets the `RecordSelectionFormula` of the
`reportClientDocument` accordingly.